finitely free modules and submodules
