+++ rings/src/linear/mod.rs
@@ -1,5 +1,7 @@
-pub mod finitely_free_affine;
-pub mod finitely_free_coset;
+pub mod finitely_free_affine_subsets;
+pub mod finitely_free_cosets;
 pub mod finitely_free_module;
+pub mod finitely_free_module_linear_transformation;
 pub mod finitely_free_submodule;
+pub mod finitely_free_submodules;
 pub mod ordered_set_free_module;
