+++ rings/src/matrix/jordan_normal_form.rs
@@ -1,7 +1,7 @@
 use super::*;
 use crate::linear::{
     finitely_free_module::FinitelyFreeModuleStructure,
-    finitely_free_submodule::FinitelyFreeSubmodule,
+    finitely_free_submodules::FinitelyFreeSubmodule,
 };
 use algebraeon_sets::sets::EnumeratedFiniteSetStructure;
 use algebraeon_structures::*;
